Update upnp discovery handling to be hue api like.
